bring out class and util-functions to separate hpp files
