+++ simulation/traffic_simulator/test/src/entity/test_entity_base.cpp
@@ -21,194 +21,8 @@
 
 #include "../catalogs.hpp"
 #include "../expect_eq_macros.hpp"
-
-class DummyEntity : public traffic_simulator::entity::EntityBase
-{
-public:
-  explicit DummyEntity(
-    const std::string & name, const traffic_simulator::CanonicalizedEntityStatus & entity_status,
-    const std::shared_ptr<hdmap_utils::HdMapUtils> & hdmap_utils_ptr)
-  : EntityBase(name, entity_status, hdmap_utils_ptr)
-  {
-  }
-
-  auto getBehaviorParameter() const -> traffic_simulator_msgs::msg::BehaviorParameter override
-  {
-    return behavior_parameter;
-  }
-
-  void setBehaviorParameter(const traffic_simulator_msgs::msg::BehaviorParameter & params) override
-  {
-    behavior_parameter = params;
-  }
-
-  auto getCurrentAction() const -> std::string override { return {}; }
-
-  auto getDefaultDynamicConstraints() const
-    -> const traffic_simulator_msgs::msg::DynamicConstraints & override
-  {
-    static const auto default_dynamic_constraints = []() {
-      auto dynamic_constraints = traffic_simulator_msgs::msg::DynamicConstraints();
-      dynamic_constraints.max_speed = 3.0;
-      dynamic_constraints.max_acceleration = 5.0;
-      dynamic_constraints.max_acceleration_rate = 7.0;
-      dynamic_constraints.max_deceleration = 11.0;
-      dynamic_constraints.max_deceleration_rate = 13.0;
-      return dynamic_constraints;
-    }();
-
-    return default_dynamic_constraints;
-  }
-
-  auto getEntityType() const -> const traffic_simulator_msgs::msg::EntityType & override
-  {
-    return entity_type;
-  }
-
-  auto getEntityTypename() const -> const std::string & override
-  {
-    static const auto str = std::string("dummy_entity");
-    return str;
-  }
-
-  ~DummyEntity() override = default;
-
-  auto getGoalPoses() -> std::vector<traffic_simulator::CanonicalizedLaneletPose> override
-  {
-    return {};
-  }
-
-  std::optional<traffic_simulator_msgs::msg::Obstacle> getObstacle() override
-  {
-    return std::nullopt;
-  }
-
-  auto getRouteLanelets(double) -> lanelet::Ids override { return std::vector<lanelet::Id>{}; }
-
-  auto getWaypoints() -> const traffic_simulator_msgs::msg::WaypointsArray override
-  {
-    return traffic_simulator_msgs::msg::WaypointsArray();
-  }
-
-  // clang-format off
-  void fillLaneletPose(traffic_simulator::CanonicalizedEntityStatus &) override {}
-  void setVelocityLimit(double) override {}
-  void setAccelerationLimit(double) override {}
-  void setAccelerationRateLimit(double) override {}
-  void setDecelerationLimit(double) override {}
-  void setDecelerationRateLimit(double) override {}
-  void requestAssignRoute(const std::vector<traffic_simulator::CanonicalizedLaneletPose> &) override {}
-  void requestAssignRoute(const std::vector<geometry_msgs::msg::Pose> &) override {}
-  void requestAcquirePosition(const traffic_simulator::CanonicalizedLaneletPose &) override {}
-  void requestAcquirePosition(const geometry_msgs::msg::Pose &) override {}
-  // clang-format on
-
-  traffic_simulator_msgs::msg::BehaviorParameter behavior_parameter;
-  traffic_simulator_msgs::msg::EntityType entity_type;
-
-  void setEntityType(uint8_t value) { entity_type.type = value; }
-
-  void appendToJobList(
-    const std::function<bool(const double)> & func_on_update,
-    const std::function<void()> & func_on_cleanup, traffic_simulator::job::Type type,
-    bool exclusive, const traffic_simulator::job::Event event)
-  {
-    job_list_.append(func_on_update, func_on_cleanup, type, exclusive, event);
-  }
-
-  std::optional<double> getTargetSpeed() { return target_speed_; }
-};
-
-auto makeHdMapUtilsSharedPointer() -> std::shared_ptr<hdmap_utils::HdMapUtils>
-{
-  std::string path =
-    ament_index_cpp::get_package_share_directory("traffic_simulator") + "/map/lanelet2_map.osm";
-  geographic_msgs::msg::GeoPoint origin;
-  origin.latitude = 35.9037067912303;
-  origin.longitude = 139.9337945139059;
-  return std::make_shared<hdmap_utils::HdMapUtils>(path, origin);
-}
-
-auto makeCanonicalizedLaneletPose(
-  std::shared_ptr<hdmap_utils::HdMapUtils> hdmap_utils, const lanelet::Id id = 120659,
-  const double s = 0.0, const double offset = 0.0)
-  -> traffic_simulator::lanelet_pose::CanonicalizedLaneletPose
-{
-  return traffic_simulator::lanelet_pose::CanonicalizedLaneletPose(
-    traffic_simulator::helper::constructLaneletPose(id, s, offset), hdmap_utils);
-}
-
-auto makeBoundingBox(const double center_y = 0.0) -> traffic_simulator_msgs::msg::BoundingBox
-{
-  traffic_simulator_msgs::msg::BoundingBox bbox;
-  bbox.center.x = 1.0;
-  bbox.center.y = center_y;
-  bbox.dimensions.x = 4.0;
-  bbox.dimensions.y = 2.0;
-  bbox.dimensions.z = 1.5;
-  return bbox;
-}
-
-auto makeEntityStatus(
-  std::shared_ptr<hdmap_utils::HdMapUtils> hdmap_utils,
-  traffic_simulator::lanelet_pose::CanonicalizedLaneletPose pose,
-  traffic_simulator_msgs::msg::BoundingBox bbox, const double speed = 0.0,
-  const std::string name = "dummy_entity") -> traffic_simulator::EntityStatus
-{
-  traffic_simulator::EntityStatus entity_status;
-  entity_status.type.type = traffic_simulator_msgs::msg::EntityType::VEHICLE;
-  entity_status.subtype.value = traffic_simulator_msgs::msg::EntitySubtype::CAR;
-  entity_status.time = 0.0;
-  entity_status.name = name;
-  entity_status.bounding_box = bbox;
-  geometry_msgs::msg::Twist twist;
-  entity_status.action_status =
-    traffic_simulator::helper::constructActionStatus(speed, 0.0, 0.0, 0.0);
-  entity_status.lanelet_pose_valid = true;
-  entity_status.lanelet_pose = static_cast<traffic_simulator::LaneletPose>(pose);
-  entity_status.pose = hdmap_utils->toMapPose(entity_status.lanelet_pose).pose;
-  return entity_status;
-}
-
-auto makeEntityStatus(
-  std::shared_ptr<hdmap_utils::HdMapUtils> /* hdmap_utils */, geometry_msgs::msg::Pose pose,
-  traffic_simulator_msgs::msg::BoundingBox bbox, const double speed = 0.0,
-  const std::string name = "dummy_entity") -> traffic_simulator::EntityStatus
-{
-  traffic_simulator::EntityStatus entity_status;
-  entity_status.type.type = traffic_simulator_msgs::msg::EntityType::VEHICLE;
-  entity_status.subtype.value = traffic_simulator_msgs::msg::EntitySubtype::CAR;
-  entity_status.time = 0.0;
-  entity_status.name = name;
-  entity_status.bounding_box = bbox;
-  geometry_msgs::msg::Twist twist;
-  entity_status.action_status =
-    traffic_simulator::helper::constructActionStatus(speed, 0.0, 0.0, 0.0);
-  entity_status.lanelet_pose_valid = false;
-  entity_status.pose = pose;
-  return entity_status;
-}
-
-auto makeCanonicalizedEntityStatus(
-  std::shared_ptr<hdmap_utils::HdMapUtils> hdmap_utils,
-  traffic_simulator::lanelet_pose::CanonicalizedLaneletPose pose,
-  traffic_simulator_msgs::msg::BoundingBox bbox, const double speed = 0.0,
-  const std::string name = "dummy_entity")
-  -> traffic_simulator::entity_status::CanonicalizedEntityStatus
-{
-  return traffic_simulator::entity_status::CanonicalizedEntityStatus(
-    makeEntityStatus(hdmap_utils, pose, bbox, speed, name), hdmap_utils);
-}
-
-auto makeCanonicalizedEntityStatus(
-  std::shared_ptr<hdmap_utils::HdMapUtils> hdmap_utils, geometry_msgs::msg::Pose pose,
-  traffic_simulator_msgs::msg::BoundingBox bbox, const double speed = 0.0,
-  const std::string name = "dummy_entity")
-  -> traffic_simulator::entity_status::CanonicalizedEntityStatus
-{
-  return traffic_simulator::entity_status::CanonicalizedEntityStatus(
-    makeEntityStatus(hdmap_utils, pose, bbox, speed, name), hdmap_utils);
-}
+#include "dummy_entity.hpp"
+#include "entity_helper_functions.hpp"
 
 int main(int argc, char ** argv)
 {
